rename fake-lms.js
