+++ components/semester-filter.js
@@ -1,7 +1,7 @@
 import './dropdown-filter';
 
 import { html, LitElement } from 'lit-element';
-import { fetchSemesters as fetchDemoSemesters } from '../model/fake-lms';
+import { fetchSemesters as fetchDemoSemesters } from '../model/fake-dataApiClient';
 import { fetchSemesters } from '../model/dataApiClient';
 import { Localizer } from '../locales/localizer';
