+++ components/user-drill-view.js
@@ -17,7 +17,7 @@ import { css, html } from 'lit-element/lit-element.js';
 import { RECORD, USER } from '../consts';
 import { createComposeEmailPopup } from './email-integration';
 import { ExportData } from '../model/exportData';
-import { fetchUserData as fetchDemoUserData } from '../model/fake-lms.js';
+import { fetchUserData as fetchDemoUserData } from '../model/fake-dataApiClient.js';
 import { fetchUserData } from '../model/dataApiClient.js';
 import { formatPercent } from '@brightspace-ui/intl';
 import { Localizer } from '../locales/localizer';
