+++ demo/index.html
@@ -9,7 +9,7 @@
 			import '@brightspace-ui/core/components/demo/demo-page.js';
 			import '../engagement-dashboard.js';
 			import fetchMock from 'fetch-mock/esm/client';
-			import { parseHash } from '../model/fake-lms';
+			import { parseHash } from '../model/fake-dataApiClient';
 
 			// setup fetchMock to print network comunication for demo purposes
 			const fetch = fetchMock.sandbox();
