+++ engagement-dashboard.js
@@ -33,7 +33,7 @@ import { Data } from './model/data.js';
 import { DiscussionActivityFilter } from './components/discussion-activity-card';
 import { ExportData } from './model/exportData';
 import { fetchData } from './model/dataApiClient.js';
-import { fetchData as fetchDemoData } from './model/fake-lms.js';
+import { fetchData as fetchDemoData } from './model/fake-dataApiClient.js';
 import { FilteredData } from './model/filteredData';
 import { heading3Styles } from '@brightspace-ui/core/components/typography/styles';
 import { Localizer } from './locales/localizer';
